@@ -5,33 +5,46 @@ if (!process.env.PWD) {
   process.env.PWD = process.cwd();
 }
 
-const output = fs.createWriteStream(process.env.PWD + '/boilerplate.zip');
-const archive = archiver('zip', {
-  zlib: { level: 9 }, // Sets the compression level.
+const builds = [
+	{
+		zip: "boilerplate.zip",
+		html: "index.html",
+		css: "css/styles.css",
+		js: "js/main.js" 
+	},
+	{
+		zip: "boilerplate-mst.zip",
+		html: "index.html.mst",
+		css: "css/styles.css",
+		js: "js/main-legacy.js"
+	}
+];
+
+builds.map(build => {
+	const stream = fs.createWriteStream(`${process.env.PWD}/${build.zip}`);
+
+	const archive = archiver('zip', {
+		zlib: { level: 9 }, // Sets the compression level.
+	});
+
+	// listen for all archive data to be written
+	// 'close' event is fired only when a file descriptor is involved
+	stream.on('close', () => {
+		console.log(`${archive.pointer()} total bytes`);
+		console.log('archiver has been finalized and the output file descriptor has closed.');
+	});
+
+	// good practice to catch this error explicitly
+	archive.on('error', (err) => {
+		throw err;
+	});
+
+	archive.pipe(stream);
+	
+	archive.append(fs.createReadStream(`${process.env.PWD}/${build.html}`), { name: build.html });
+	archive.append(fs.createReadStream(`${process.env.PWD}/${build.css}`), { name: build.css });
+	// Ensure the outputted JS has the same name the html file is looking for
+	archive.append(fs.createReadStream(`${process.env.PWD}/${build.js}`), { name: "js/main.js" });
+	
+	archive.finalize();
 });
-
-// listen for all archive data to be written
-// 'close' event is fired only when a file descriptor is involved
-output.on('close', () => {
-  console.log(`${archive.pointer()} total bytes`);
-  console.log('archiver has been finalized and the output file descriptor has closed.');
-});
-
-// good practice to catch this error explicitly
-archive.on('error', (err) => {
-  throw err;
-});
-
-archive.pipe(output);
-
-// append a file from stream
-const file1 = `${process.env.PWD}/js/main.js`;
-archive.append(fs.createReadStream(file1), { name: '/js/main.js' });
-
-const file2 = `${process.env.PWD}/css/styles.css`;
-archive.append(fs.createReadStream(file2), { name: '/css/styles.css' });
-
-const file3 = `${process.env.PWD}/index.html.mst`;
-archive.append(fs.createReadStream(file3), { name: 'index.html.mst' });
-
-archive.finalize();
