@@ -1,30 +1,36 @@
 // Please put all config needed for the tempate into the array bellow
 const template = new boilerplate({
   fonts: ['PUT_ALL_FONT_NAMES_HERE'],
-  trimMarks: templateProps['trim-marks'] === 'true',
-  cssVariables,
-  templateProps,
   hotReloadOnChange: true,
-  placeholderVisibility: templateProps['placeholder-visibility'],
-  placeholderImages: templateProps['placeholder-1'],
-});
-
-// This function will run whenever there is a change to any text input, resize and on inital load
-template.on('inputs-change', async (inputs) => {
-  console.log(inputs);
-
-  document.querySelector('h1').innerHTML = `Hello, ${inputs.user.name}!<br>I appreciate you`;
-  template.textFit(document.querySelectorAll('h1'), { fontUnit: 'rem', minFontSize: 0.5, maxFontSize: 1.5 });
-
-  await template.ensureAllImagesLoaded();
-
-  // template.dynamicReplace();
-  // template.maxHeightCheck();
-  // template.maxLineCheck();
-  // template.charLimit();
-  template.completeRender();
 });
 
 // template.on('overflow', () => {
 //   console.log(template.overflows)
 // })
+
+// This function will run whenever there is a change to any text input, resize and on inital load
+template.on('inputs-change', async (inputs) => {
+  const { account, team, user } = inputs;
+  console.log(inputs);
+
+  template.trim = inputs.trimMarks;
+  template.setupPlaceholder(inputs['placeholder-1']); // remove this line if you are not using a placeholder image
+  // This is where you can do all your magic
+  document.querySelectorAll('h1').forEach((element) => {
+    element.innerText = `Welcome to ${account.name}!`;
+  });
+  template.textFit(document.querySelectorAll('h1'), {
+    fontUnit: '%',
+    minFontSize: 50,
+    maxFontSize: 100,
+    widthOnly: true,
+    maxLine: 1,
+  }); // this code will use textfit to keep the element on a single line or overflow if it can't at half the font-size
+
+  await template.ensureAllImagesLoaded();
+  template.maxHeightCheck(document.querySelectorAll('.container'));
+  template.maxLineCheck(document.querySelectorAll('.body-copy'), 4);
+  // template.dynamicReplace();
+  // template.charLimit();
+  template.completeRender();
+});
