@@ -1,15 +1,16 @@
 // Webpack uses this to work with directories
 const path = require('path');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 // This is the main configuration object.
 // Here, you write different options and tell Webpack what to do
 module.exports = {
   // Default mode for Webpack is production.
   // Depending on mode Webpack will apply different things
-  // on the final bundle. For now, we don't need production's JavaScript 
+  // on the final bundle. For now, we don't need production's JavaScript
   // minifying and other things, so let's set mode to development
   mode: 'development',
-  devtool: 'inline-source-map',
+  devtool: 'source-map',
   // Path to your entry point. From this file Webpack will begin its work
   entry: './modules/index.js',
   // Path and filename of your result bundle.
@@ -29,6 +30,37 @@ module.exports = {
         exclude: [/node_modules/],
         use: ['babel-loader'],
       },
+      {
+        test: /\.(css|scss)$/,
+        use: [
+          {
+            loader: MiniCssExtractPlugin.loader,
+          },
+          {
+            // Interprets `@import` and `url()` like `import/require()` and will resolve them
+            loader: 'css-loader',
+          },
+          {
+            // Loader for webpack to process CSS with PostCSS
+            loader: 'postcss-loader',
+            options: {
+              postcssOptions: {
+                plugins: [['autoprefixer']],
+              },
+            },
+          },
+          {
+            // Loads a SASS/SCSS file and compiles it to CSS
+            loader: 'sass-loader',
+          },
+        ],
+      },
     ],
-  }
+  },
+  plugins: [
+    new MiniCssExtractPlugin({
+      // Options similar to the same options in webpackOptions.output
+      filename: 'main.css',
+    }),
+  ],
 };
